+++ controllers/operator/mongodbopsmanager_controller_test.go
@@ -303,11 +303,14 @@ func TestOpsManagerCentralUrl(t *testing.T) {
 
 // TODO move this test to 'opsmanager_types_test.go' when the builder is moved to 'apis' package
 func TestOpsManagerBackupDaemonHostName(t *testing.T) {
-	assert.Equal(t, "test-om-backup-daemon-0",
-		DefaultOpsManagerBuilder().Build().BackupDaemonHostName())
+	assert.Equal(t, []string{"test-om-backup-daemon-0"},
+		DefaultOpsManagerBuilder().Build().BackupDaemonHostNames())
 	// The host name doesn't depend on cluster domain
-	assert.Equal(t, "test-om-backup-daemon-0",
-		DefaultOpsManagerBuilder().SetClusterDomain("some.domain").Build().BackupDaemonHostName())
+	assert.Equal(t, []string{"test-om-backup-daemon-0"},
+		DefaultOpsManagerBuilder().SetClusterDomain("some.domain").Build().BackupDaemonHostNames())
+
+	assert.Equal(t, []string{"test-om-backup-daemon-0", "test-om-backup-daemon-1", "test-om-backup-daemon-2"},
+		DefaultOpsManagerBuilder().SetBackupMembers(3).Build().BackupDaemonHostNames())
 }
 
 func TestTriggerOmChangedEventIfNeeded(t *testing.T) {