+++ docker/mongodb-enterprise-tests/tests/opsmanager/om_jvm_params.py
@@ -59,7 +59,9 @@ class TestOpsManagerCreationWithJvmParams:
         assert int(rss) / 1024 > 400
 
     def test_om_jvm_backup_params_configured(self, ops_manager: MongoDBOpsManager):
-        pod_name = ops_manager.backup_daemon_pod_name()
+        pod_names = ops_manager.backup_daemon_pods_names()
+        assert len(pod_names) == 1
+        pod_name = pod_names[0]
         cmd = ["/bin/sh", "-c", "cat " + OM_CONF_PATH_DIR]
 
         result = KubernetesTester.run_command_in_pod_container(