+++ docker/mongodb-enterprise-tests/tests/opsmanager/om_ops_manager_backup_light.py
@@ -116,7 +116,8 @@ class TestOpsManagerCreation:
     ):
         om_tester = ops_manager.get_om_tester()
         om_tester.assert_healthiness()
-        om_tester.assert_daemon_enabled(ops_manager.backup_daemon_pod_name(), HEAD_PATH)
+        for pod_name in ops_manager.backup_daemon_pods_names():
+            om_tester.assert_daemon_enabled(pod_name, HEAD_PATH)
         om_tester.assert_oplog_stores([new_om_data_store(oplog_replica_set, "oplog1")])
 
         ops_manager.appdb_status().assert_reaches_phase(Phase.Running, timeout=600)