+++ docker/mongodb-enterprise-tests/tests/opsmanager/om_ops_manager_backup_manual.py
@@ -258,7 +258,9 @@ class TestOpsManagerCreation:
     def test_om(self, ops_manager: MongoDBOpsManager):
         om_tester = ops_manager.get_om_tester()
         om_tester.assert_healthiness()
-        om_tester.assert_daemon_enabled(ops_manager.backup_daemon_pod_name(), HEAD_PATH)
+        for pod_name in ops_manager.backup_daemon_pods_names():
+            om_tester.assert_daemon_enabled(pod_name, HEAD_PATH)
+
         # No oplog stores were created in Ops Manager by this time
         om_tester.assert_oplog_stores([])
         om_tester.assert_s3_stores([])